Main restructured to modules
