+++ constants.py
@@ -2,3 +2,4 @@ WINDOW_WIDTH = 900
 WINDOW_HEIGHT = 600
 MAIN_CHARACTER_SCALE = 1.5
 TICKRATE = 1 / 120.0
+WALKING_SPEED = 200
